move xdg-open to proper paths
do not put xdg-open into /usr/local. we kept it there initially with the possibility of a personal-core snap in mind where the real xdg-open would be included but this is unlikely to happen. instead of having to mangle the PATH variable for snap wrappers the binary can now be found right away in $PATH without extra adjustment.