Fix handling of secondary addresses
The default behaviour of the kernel is to take down the secondary
addresses when a primary address goes away. However this is not
what we want with networkd (and the systemd upstream changes this
in its sysctl.d/50-default.conf in the upstream git tree). The
way networkd works is to add a new DHCP IP as a secondary address
and when the previous address goes away the secondary address
needs to be promoted, not removed. This patch fixes this behaviour.

For reference:

Systemd sysctl defaults:
- https://github.com/systemd/systemd/blob/master/sysctl.d/50-default.conf

Systemd-networkd dhcp renew address failure with all the details:
- https://bugs.launchpad.net/bugs/1721223