live-build: fix order of -print0 in find
The order of find is critical for it to work correctly. Each item
after the starting point is evaluated and the resulting return is
used to consider if the term should continue or not. This means
that: "find ./foo -print0 -type f" will *always* call -print0
(and return an implicit "true") and then move on to the "-type f"
test which will return false and the processing will stop for
the given file. So to make things work correctly we need to
write it in the right order as find just evaluates expressions.