hooks: make sure the motd 'welcome' string is build based on uname -a
This commit leverages the update-motd script and changes the static
motd in a way that the 'Welcome to Ubuntu' string is now generated
by the update-motd. The direct benefit is that it is build using
uname -a and such therefore it contains a better information.
Moreover, future changes may add system measurements here.