hooks: create compat copy of /lib/udev/snappy-app-dev
This creates a compatibility copy /usr/lib/snapd/snap-device-helper
to /lib/udev/snappy-app-dev so that an old snap-confine will work
correctly with a new core snap. The snappy-app-dev binary got
renamed and while we provided a symlink this is not sufficient
because the apparmor profile for snap-confine does only allow
running /lib/udev/snappy-app-dev.