hooks: unwind /etc/alternatives
The core snap currently ships with a bunch of files pointing to
/etc/alternatives. E.g.

    /usr/bin/awk -> /etc/alternatives/awk -> /usr/bin/gawk

This is problematic because /etc is mounted from the host so
a user how configured a different awk which is not part of
core will have a dangling symlink (because /usr comes from
the core snap of course).

There is a small risk with this PR as it changes the existing
behaviour. However I would argue its worth it because the
current "awk" that you get in the core snap depends on the
host you run your snap in. Which is not the predictable
behaviour we want.