hooks: update motd handling and remove 50-motd-news
The base-files package grew a new /etc/update-motd.d/50-motd-news
script during the lifetime of Core 16. This script wants to write
to the non-writable /var/cache/motd-news location and also needs
curl to work. So it fails and that leads to breaking the
/etc/update-motd.d/00-header script.

This commit disables the script and also adds a check to ensure
that no new files are added to /etc/update-motd.d without us
noticing.