hooks: replaces symlinks in /usr/share/doc with the actual files
We see a lot of issues in the review-tools because the symlinks
in /usr/share/doc for NEWS.Debian.gz and changelog.Debian.gz are
not predictable. This commit makes them predictable by replacing
the links with the actual file. This is fine because the data
is stored inside a squashfs image so the duplication of the
strings will get (mostly) compressed away. Plus the files are
small anyway.

This should eliminate the need for the manual reviews of the
core snap.