hooks: remove "fuse" from the core snap
With the introduction of https://github.com/snapcore/snapd/pull/11600
now the `fuse` package is installed on `core`. However we do not
want this. This commit forcefully removes it if it gets installed.

There is also a followup https://github.com/snapcore/snapd/pull/11648
that will downgrade fuse to a recommends.