hooks: restore link to built initrd
Recent changes in xenial livebuild include the fuse package when
creating the chroot - this was not until very recently. When it gets
removed by hook 600-no-debian.binary, the initrd gets re-built and the
symlink created by 25-create-generic-initrd.chroot is replaced by a
real file. Put things back to normal by making it point back to the
initrd generated by the latter.